Training_center (spring)/src/main/java/com/krylova/service/CompanyService.java
package com.krylova.service;

import com.krylova.entity.Company;
import com.krylova.repository.CompanyRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class CompanyService {

    @Autowired
    private CompanyRepository companyRepository;

    public Company create(Company company) {
        return companyRepository.save(company);
    }

    public Company update(Company company) {
        return companyRepository.save(company);
    }

    public boolean delete(Company company) {
        companyRepository.delete(company);
        return true;
    }

    public List<Company> findAll() {
        return companyRepository.findAll();
    }

    public Optional<Company> findById(Long id) {
        return companyRepository.findById(id);
    }

    public List<Company> findByName(String name) {
        return companyRepository.findByName(name);
    }

}
